Parser

Documentation

parseHeader :: CppHeader a => ParserGraph a -> Parser a

buildAst :: CppHeader a => ParserGraph a -> [CppToken] -> String -> Either (ParseErrorBundle [CppToken] Void) a

tokenToString :: CppToken -> String

tokensToString :: [CppToken] -> String

punctuator :: Char -> Parser Char

keyword :: String -> Parser String

identifier :: Parser String

tag :: String -> Parser String

matchIdentifier :: String -> Parser String

qualifiedName :: Parser String

Namespace Grammar

namespaceDefinition :: CppNamespace a => String -> ParserGraph a -> Parser a

BNF Grammar:

namespace-definition:	 
 	named-namespace-definition
 	unnamed-namespace-definition

See https://alx71hub.github.io/hcb/#namespace.def for more context

namedNamespaceDefinition :: CppNamespace a => String -> ParserGraph a -> Parser a

BNF Grammar:

named-namespace-definition:	 
 	original-namespace-definition
 	extension-namespace-definition

original-namespace-definition:	 
 	inline (opt) namespace identifier { namespace-body }

extension-namespace-definition:	 
 	inline (opt) namespace original-namespace-name { namespace-body }

Notes:

original-namespace-name = identifier

namespace-body is parsed by other parsers injected via the ParserGraph

See https://alx71hub.github.io/hcb/#namespace.def for more context

unnamedNamespaceDefinition :: CppNamespace a => String -> ParserGraph a -> Parser a

unnamed-namespace-definition:	 
 	inline (opt) namespace { namespace-body }

Notes:

original-namespace-name = identifier

namespace-body is parsed by other parsers injected via the ParserGraph

See https://alx71hub.github.io/hcb/#namespace.def for more context

Enum Grammar

enumSpecifier :: CppEnum a => String -> ParserGraph a -> Parser a

BNF Grammar:

enum-specifier:	 
 	enum-head { enumerator-list (opt) }
 	enum-head { enumerator-list , }

Notes:

enumerator-list is parsed by other parsers injected via the ParserGraph

See https://alx71hub.github.io/hcb/#dcl.enum for more context

enumHead :: Parser CppEnumHead

BNF Grammar:

enum-head:	 
 	enum-key attribute-specifier-seq (opt) identifier (opt) enum-base (opt)
 	enum-key attribute-specifier-seq (opt) nested-name-specifier identifier enum-base (opt)

See https://alx71hub.github.io/hcb/#dcl.enum for more context

enumKey :: Parser ()

BNF Grammar:

enum-key:	 
 	enum
 	enum class
 	enum struct

See https://alx71hub.github.io/hcb/#dcl.enum for more context

enumBase :: Parser [CppTypeSpecifier]

BNF Grammar:

enumerator-definition:	 
 	enumerator
 	enumerator = constant-expression

See https://alx71hub.github.io/hcb/#dcl.enum for more context

enumeratorDefinition :: CppEnumerator a => Parser a

Class Grammar

classSpecifier :: CppClass a => String -> ParserGraph a -> Parser a

BNF Grammar:

class-specifier:	 
 	class-head { member-specification (opt) }

Notes:

member-specification is parsed by other parsers injected via the ParserGraph

See https://alx71hub.github.io/hcb/#class for more context

classHead :: Parser CppClassHead

BNF Grammar:

class-head:	 
 	class-key attribute-specifier-seq (opt) class-head-name class-virt-specifier-seq (opt) base-clause (opt)
 	class-key attribute-specifier-seq (opt) base-clause (opt)

See https://alx71hub.github.io/hcb/#class for more context

classKey :: Parser ()

BNF Grammar:

class-key:	 
 	class
 	struct
 	union

See https://alx71hub.github.io/hcb/#class for more context

classVirtSpecifierSeq :: Parser [String]

BNF Grammar:

class-virt-specifier-seq:	 
 	class-virt-specifier
 	class-virt-specifier-seq class-virt-specifier  

Notes:

Grammar for for class-virt-specifier-seq is recursive and therfore can be sumarised to some ... in haskell

See https://alx71hub.github.io/hcb/#class for more context

classVirtSpecifier :: Parser String

BNF Grammar:

class-virt-specifier:	 
 	final

See https://alx71hub.github.io/hcb/#class for more context

baseClause :: Parser [CppBaseSpecifier]

baseSpecifierList :: Parser [CppBaseSpecifier]

BNF Grammar:

base-specifier-list:	 
 	base-specifier ... (opt)
 	base-specifier-list , base-specifier ... (opt)

Notes:

Grammar for for base-specifier-list is recursive and therfore can be sumarised to some ... in haskell

See https://alx71hub.github.io/hcb/#class.derived for more context

baseSpecifier :: Parser CppBaseSpecifier

accessSpecifier :: Parser CppAccessSpecifier

BNF Grammar:

access-specifier:	 
 	private
 	protected
 	public

See https://alx71hub.github.io/hcb/#class.derived for more context

classOrDeclType :: Parser String

BNF Grammar:

class-or-decltype:	 
 	:: (opt) nested-name-specifieropt class-name
 	decltype-specifier

Notes:

Currently decltype-specifier is not accepted as a valid class-or-decltype

TODO : Support decltype-specifier as a valid class-or-decltype

See https://alx71hub.github.io/hcb/#class.derived for more context

className :: Parser String

BNF Grammar:

class-name:	 
 	identifier
 	simple-template-id 

See https://alx71hub.github.io/hcb/#class for more context

Attribute Specifier Grammar

attributeSpecifierSeq :: Parser [CppAttributeSpecifier]

BNF Grammar:

attribute-specifier-seq:	 
 	attribute-specifier
 	attribute-specifier-seq attribute-specifier

Notes:

Grammar for for attribute-specifier-seq is recursive and therfore can be sumarised to some ... in haskell

See https://alx71hub.github.io/hcb/#dcl.attr.grammar for more context

attributeSpecifier :: Parser CppAttributeSpecifier

BNF Grammar:

attribute-specifier:	 
 	[ [ attribute-list ] ]
 	alignment-specifier

alignment-specifier:	 
 	alignas ( type-id ... (opt) )
 	alignas ( assignment-expression ... (opt) )

Notes:

attribute-specifier is parsed as "token soup" (an array of tokens, how you deal with this is up to you)

See https://alx71hub.github.io/hcb/#dcl.attr.grammar for more context

Type Specifier Grammar

typeSpecifierSeq :: Parser [CppTypeSpecifier]

BNF Grammar:

type-specifier-seq:	 
 	type-specifier attribute-specifier-seq (opt)
 	type-specifier type-specifier-seq

Notes:

Grammar for for type-specifier-seq is recursive and therfore can be sumarised to some ... in haskell

See https://alx71hub.github.io/hcb/#dcl.type for more context

typeSpecifier :: Parser String

BNF Grammar:

type-specifier:	 
 	trailing-type-specifier
 	class-specifier
 	enum-specifier

Notes:

Currently class-specifier and enum-specifier are not accepted as a valid type-specifier

TODO : Support class-specifier and enum-specifier as a valid type-specifier

See https://alx71hub.github.io/hcb/#dcl.type for more context

trailingTypeSpecifier :: Parser String

BNF Grammar:

trailing-type-specifier:	 
 	simple-type-specifier
 	elaborated-type-specifier
 	typename-specifier
 	cv-qualifier

Notes:

Currently elaborated-type-specifier and typename-specifier are not accepted as a valid trailing-type-specifier

TODO : Support elaborated-type-specifier and typename-specifier as a valid trailing-type-specifier

See https://alx71hub.github.io/hcb/#dcl.type for more context

simpleTypeSpecifier :: Parser String

BNF Grammar:

simple-type-specifier:	 
 	:: (opt) nested-name-specifier (opt) type-name
 	:: (opt) nested-name-specifier template simple-template-id
 	char
 	char16_t
 	char32_t
 	wchar_t
 	bool
 	short
 	int
 	long
 	signed
 	unsigned
 	float
 	double
 	void
 	auto
 	decltype-specifier

Notes:

nested-name-specifier (opt) type-name is handled via the function typeName

typeName uses the function qualifiedName which internally implements nested-name-specifier grammar

Currently decltype-specifier is not accepted as a valid trailing-type-specifier

TODO : Support decltype-specifie as a valid simple-type-specifier

TODO : support optional preceeding :: for :: (opt) nested-name-specifier (opt) type-name and :: (opt) nested-name-specifier template simple-template-id

See https://alx71hub.github.io/hcb/#dct.type.simple for more context

typeName :: Parser String

BNF Grammar:

type-name:	 
 	class-name
 	enum-name
 	typedef-name
 	simple-template-id

Notes:

enum-name = identifier , typedef-name = identifier , class-name = identifier or simple-template-id

From this we can see that a type-name is either an identifier or simple template-id

Everywhere type-name occurs it is preceded by nested-name-specifier so lets just use qualifiedName instead of identifier

See https://alx71hub.github.io/hcb/#dct.type.simple for more contex

simpleTemplateID :: Parser String

BNF Grammar:

simple-template-id:	 
 	template-name (opt)

Notes:

template-name = identifier

simple-template-id is parsed as a single string, for example vectorint will be parsed completely as "vectorint"

See https://alx71hub.github.io/hcb/#dct.type.simple for more contex

cvQualifier :: Parser String

BNF Grammar:

cv-qualifier:	 
 	const
 	volatile

See https://alx71hub.github.io/hcb/#dcl.decl for more contex

Class Member Grammar

simpleVariableDeclaration :: CppMemberVariable a => Parser a